Displaying all figures & Documentation changes
I display all the figures available to the players and made them move
right or down, just to test if they are well defined. I also intorduced
some changes into the Specification: the board division is no longer
displayed, the figures are numbered from 1 to 7 and I corrected some
spelling errors. Present verion of the specification is version 1.01.