Change in Next Figure display + fixed glitched
Instead of drawing figures I use 7 bitmaps of the section with a figure
already included, one for each figure. I load one bitmap. When I want to
change the display, I delete the first bitmap and then load another one.
I fixed the flickering of the display. It doesn't go back to the
original. I moved the display code before the while loop.